Cleanup extforms/workshops forms
1. Moved TrainingRequestForm to extforms.forms
2. Moved field definitions from workshops.forms to workshops.fields
3. Adjusted styling in some files
4. Joined extremely similar tests against TrainingRequestForm
