+++ amy/workshops/tests/test_training_request.py
@@ -1,7 +1,6 @@
 import unittest
 from urllib.parse import urlencode
 
-from django.core import mail
 from django.core.exceptions import ValidationError
 from django.template import Context
 from django.template import Template
@@ -18,63 +17,6 @@ from workshops.models import (
     Task,
     KnowledgeDomain,
 )
-
-
-class TestTrainingRequestForm(TestBase):
-    def setUp(self):
-        self._setUpUsersAndLogin()
-        self._setUpRoles()
-
-    def test_request_added(self):
-        email = 'john@smith.com'
-        data = {
-            'group': 'coolprogrammers',
-            'personal': 'John',
-            'family': 'Smith',
-            'email': email,
-            'github': '',
-            'occupation': '',
-            'occupation_other': 'unemployed',
-            'affiliation': 'AGH University of Science and Technology',
-            'location': 'Cracow',
-            'country': 'PL',
-            'domains': [1, 2],
-            'domains_other': '',
-            'underrepresented': '',
-            'previous_involvement': [Role.objects.get(name='host').id],
-            'previous_training': 'none',
-            'previous_training_other': '',
-            'previous_training_explanation': '',
-            'previous_experience': 'none',
-            'previous_experience_other': '',
-            'previous_experience_explanation': '',
-            'programming_language_usage_frequency': 'daily',
-            'reason': 'Just for fun.',
-            'teaching_frequency_expectation': 'monthly',
-            'teaching_frequency_expectation_other': '',
-            'max_travelling_frequency': 'yearly',
-            'max_travelling_frequency_other': '',
-            'addition_skills': '',
-            'comment': '',
-            'data_privacy_agreement': 'on',
-            'code_of_conduct_agreement': 'on',
-            'training_completion_agreement': 'on',
-            'workshop_teaching_agreement': True,
-        }
-        self.passCaptcha(data)
-
-        rv = self.client.post(reverse('training_request'), data,
-                              follow=True)
-        self.assertEqual(rv.status_code, 200)
-        content = rv.content.decode('utf-8')
-        self._save_html(content)
-        self.assertNotIn('Fix errors below', content)
-        self.assertEqual(TrainingRequest.objects.all().count(), 1)
-
-        # Test that the sender was emailed
-        self.assertEqual(len(mail.outbox), 1)
-        msg = mail.outbox[0]
-        self.assertEqual(msg.to, [email])
 
 
 def create_training_request(state, person):
